fix: Moved util/pan in namespace ol/util
