+++ src/webgis4u/ol/util/pan.test.js
