+++ src/webgis4u/util/ol.js
@@ -4,4 +4,3 @@
  */
 
 export * from './ol/getMapChildElementByClassName';
-export * from './ol/pan';
